exceptions moved to one package and createOrder method added to OrderController
