+++ src/main/java/com/kodilla/ecommercee/exception/CartNotFoundException.java
@@ -1,4 +1,4 @@
-package com.kodilla.ecommercee.controller;
+package com.kodilla.ecommercee.exception;
 
 public class CartNotFoundException extends Exception {
 }
