+++ src/main/java/com/kodilla/ecommercee/exception/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package com.kodilla.ecommercee.controller;
+package com.kodilla.ecommercee.exception;
 
 public class UserNotFoundException extends Exception {
 }
